api/requirements.txt
# FastAPI核心依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 异步支持
aiofiles==23.2.1
httpx==0.26.0

# SSE流式响应
sse-starlette==1.8.2

# 安全认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 速率限制
slowapi==0.1.9

# 现有文档处理依赖
langchain
langchain-community
langchain-core
qdrant-client
langchain-qdrant

# 嵌入模型
dashscope

# 文档加载和处理
beautifulsoup4
html2text
markitdown

# 文本分割
tiktoken
transformers

# Redis缓存（可选）
redis==5.0.1

# 日志和监控（可选）
python-json-logger==2.0.7
loguru==0.7.2

# OpenTelemetry（可选）
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0

# Prometheus指标（可选）
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.19.0

# 工具库
python-dotenv
tenacity==8.2.3

# JSON格式日志
json-log-formatter

types-python-jose
types-passlib
